+++ frontend/src/App.tsx
@@ -25,6 +25,7 @@ import Photos from './pages/Photos'
 import Notes from './pages/Notes'
 import Admin from './pages/Admin'
 import Finances from './pages/Finances'
+import WaterChangeCalculator from './pages/WaterChangeCalculator'
 import PublicTankProfile from './pages/PublicTankProfile'
 
 const Welcome = lazy(() => import('./pages/Welcome'))
@@ -77,6 +78,7 @@ function App() {
             <Route path="photos" element={<Photos />} />
             <Route path="notes" element={<Notes />} />
             <Route path="finances" element={<Finances />} />
+            <Route path="water-change-calculator" element={<WaterChangeCalculator />} />
             {!local && <Route path="admin" element={<Admin />} />}
           </Route>